Type-check memgit.core and memgit.retrieval
Two genuine gaps found and fixed along the way: diff.py's early-return
branches reused the name `values` for a tuple where the loop later in
the same function binds it as a list, which mypy read as one variable
redefined with an incompatible type -- renamed the two tuple bindings.
Retriever._vector_for's post-write cache readback was typed as
possibly None even though a fact just written is always readable back;
made that invariant an explicit assertion instead of an implicit one.
The rest is typeshed's `float ** float -> Any` overload and a
json.loads() return needing its shape asserted.

+++ src/memgit/core/decay.py
@@ -142,7 +142,11 @@ class DecayPolicy:
             return fact.confidence
 
         age = self.age_days(fact, as_of=as_of)
-        value = fact.confidence * 0.5 ** (age / half_life)
+        # `0.5 ** x` types as `Any` in typeshed -- a negative base raised to a
+        # fractional exponent can be complex, which a `float`-only overload
+        # can't express. The base here is always positive, so the result is
+        # always a real float; the explicit cast reclaims that for mypy.
+        value = fact.confidence * float(0.5 ** (age / half_life))
         return max(self.floor, value)
 
     # -- derivation ------------------------------------------------------
